Correct sign up test

+++ src/test/java/org/pesmypetcare/webservice/controller/appmanager/MyPetCareRestControllerTest.java
@@ -3,7 +3,7 @@ package org.pesmypetcare.webservice.controller.appmanager;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
 import com.google.firebase.auth.FirebaseAuthException;
-import org.junit.jupiter.api.AfterAll;
+import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 import org.pesmypetcare.webservice.error.ErrorBody;
@@ -16,6 +16,7 @@ import org.springframework.test.web.servlet.MockMvc;
 
 import static org.springframework.test.util.AssertionErrors.assertEquals;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 @SpringBootTest
 @AutoConfigureMockMvc
@@ -26,14 +27,18 @@ class MyPetCareRestControllerTest {
     private static String url;
     private static String key;
     private static String password;
+    private static boolean deleteUser1;
 
     @Autowired
     private MockMvc mockMvc;
 
-    @AfterAll
-    public static void resetFirebase() {
+    @AfterEach
+    public void resetFirebase() {
         try {
-            deleteUser("user1");
+            if (deleteUser1) {
+                deleteUser("user1");
+                deleteUser1 = false;
+            }
         } catch (FirebaseAuthException e) {
             e.printStackTrace();
         }
@@ -54,11 +59,16 @@ class MyPetCareRestControllerTest {
         url = "/signup";
         password = "123456";
         key = "password";
+        deleteUser1 = false;
     }
 
     @Test
     public void signUpShouldReturnStatusOK() throws Exception {
-        doSignIn(jsonUser1);
+        mockMvc.perform(post(url)
+            .contentType(MediaType.APPLICATION_JSON)
+            .content(jsonUser1).param(key, password))
+            .andExpect(status().isOk());
+        deleteUser1 = true;
     }
 
     @Test
@@ -68,6 +78,7 @@ class MyPetCareRestControllerTest {
         String msg = getBodyMessage(body);
         assertEquals("Response message when signing up with a registered username",
             "The specified uid already exists", msg);
+        deleteUser1 = true;
     }
 
     @Test
@@ -77,6 +88,7 @@ class MyPetCareRestControllerTest {
         String msg = getBodyMessage(body);
         assertEquals("Response message when signing up with a registered email",
             "The specified email is already in use", msg);
+        deleteUser1 = true;
     }
 
     @Test
